+++ Canale_Andrea/.gitignore
@@ -1,4 +1,2 @@
-/node_modules/
-/public/build/
-user.json
-.DS_Store
+node_modules
+.svelte-kit
