--- Canale_Andrea/src/index.svelte
@@ -1,9 +1,0 @@
-<script>
-    import { Router, Route, Link } from "svelte-navigator";
-    export let indexRoute;
-  </script>
-  
-  <div class="Blog">
-<h1>Ciao</h1>
-  </div>
-
