--- Canale_Andrea/src/main.js
@@ -1,7 +1,0 @@
-import App from './App.svelte';
-
-var app = new App({
-	target: document.body
-});
-
-export default app;
